// client/src/components/CountrySearch/Country.types.ts
export interface ICountry {
  altSpellings?: string[];
  area?: number;
  capital?: string[];
  capitalInfo?: { latlng?: number[] };
  car?: { signs?: string[]; side?: string };
  cca2?: string;
  cca3?: string;
  ccn3?: string;
  cioc?: string;
  coatOfArms?: { png?: string; svg?: string };
  continents?: string;
  currencies?: { USD?: { name?: string; symbol?: string } };
  demonyms?: { eng?: { f?: string; m?: string } };
  fifa?: string;
  flag?: string; //"🇬🇺"
  flags?: { png?: string; svg?: string };
  idd?: { root?: string; suffixes: string[] };
  independent?: boolean;
  landlocked?: boolean;
  languages?: { [key: string]: string };
  latlng?: number[];
  maps?: { googleMaps?: string; openStreetMaps?: string };
  name?: {
    common?: string;
    official?: string;
    nativeName?: { [key: string]: string };
  };
  population?: number;
  postalCode?: { format?: string; regex?: string };
  region?: string;
  startOfWeek?: string;
  status?: string;
  subregion?: string;
  timezones?: string[];
  tld?: string[];
  translations?: {
    ara?: { official?: string; common?: string };
    ces?: { official?: string; common?: string };
    cym?: { official?: string; common?: string };
    deu?: { official?: string; common?: string };
    est?: { official?: string; common?: string };
    fin?: { official?: string; common?: string };
    fra?: { official?: string; common?: string };
    hrv?: { official?: string; common?: string };
    hun?: { official?: string; common?: string };
    ita?: { official?: string; common?: string };
    jpn?: { official?: string; common?: string };
    kor?: { official?: string; common?: string };
    nld?: { official?: string; common?: string };
    per?: { official?: string; common?: string };
    pol?: { official?: string; common?: string };
    por?: { official?: string; common?: string };
    rus?: { official?: string; common?: string };
    slk?: { official?: string; common?: string };
    spa?: { official?: string; common?: string };
    swe?: { official?: string; common?: string };
    urd?: { official?: string; common?: string };
    zho?: { official?: string; common?: string };
  };
  unMember?: boolean;
}